Get info about a variable

Get info about a variable by applying the relevant varInfo entries.

.vsc.applyVarInfos(
  v,
  infos = character(0),
  stackingInfos = character(0),
  verbose = getOption("vsc.verboseVarInfos", FALSE),
  ind = NULL
)

Arguments

v

The variable, any R value.

infos

Character vector of field names in varInfo. Retrieve first match.

stackingInfos

Character vector of field names in varInfo. Retrieve all matches.

verbose

Whether to print debug info using logCat, logPrint.

ind

The indices to retrieve if child variables are retrieved.

Value

A named list, containing the corresponding varInfo results.

Details

The allowed varInfo entries and their types are:

childVars: MinimalVariable[]

nChildVars: number

customAttributes: MinmalVariable[]

internalAttributes: MinmalVariable[]

toString: string

type: string

evaluateName: string

printFunc: function | boolean

Where MinimalVariable[] refers to a list of named lists with entries:

name: string

rValue: any R value

setter: (optional, undocumented)

setInfo: (optional, undocumented)

See also